added minimal working example of 3d interpolation
